Fix transaction error
Problem:
After a transaction was commited the messages in the queue should be processed. This was not the case.

Solution:
There were two different queues. The messages were queued at server.rs, however the queue of consensus.rs was handled. I removed the queue at consensus and changed the `handle_queue` method.