Implement community_string also for ServerAdd
When a node wants to join the cluster, it must send a ServerAdd to a node which is already in the cluster. Then the node which is already in the cluster sends a ServerConnectionPreamble and expects also a reply of the new node. This means the new node receives the message with the community_string first which is a security issue. In order to fix it, the community_string has to be checked when a node gets a ServerAdd message.